show error to user when failed to get newest pool contributors
